+++ packages/hono-takibi/package.json
@@ -3,10 +3,7 @@
   "description": "Hono Takibi is a CLI tool that generates Hono routes from OpenAPI specifications.",
   "version": "0.4.2",
   "license": "MIT",
-  "keywords": [
-    "hono",
-    "openapi"
-  ],
+  "keywords": ["hono", "openapi"],
   "homepage": "https://github.com/nakita628/hono-takibi",
   "publishConfig": {
     "access": "public"
@@ -20,9 +17,7 @@
   },
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "files": ["dist"],
   "bin": {
     "hono-takibi": "dist/index.js"
   },
