+++ packages/hono-takibi/src/generators/zod-openapi-hono/app/docs/generate-docs.ts
@@ -1,4 +1,4 @@
-import type { OpenAPISpec } from '../../../types'
+import type { OpenAPISpec } from '../../../../types'
 
 export function generateDocs(openAPISpec: OpenAPISpec): {
   openapi: string | undefined
