+++ packages/hono-takibi/src/generators/zod-openapi-hono/openapi/route/params/generate-params-object.ts
@@ -1,4 +1,3 @@
-
 import type { Parameters, ParamsObject } from '../../../../../types'
 import type { Config } from '../../../../../config'
 import { generateZodCoerce } from '../../../../zod/generate-zod-coerce'
